+++ src/promoters/PromotePackageJson/PromotePackageJson.tests.ts
@@ -79,6 +79,49 @@ describe.each`
 
 describe.each`
 	currentVersion           | nextVersion
+	${"2.0.0"}               | ${"2.1.0"}
+	${"5.7.1-rc.0+93f9b843"} | ${"5.7.1-rc.0"}
+`(
+	"when the package.json file has a 'version' field of $currentVersion preceded by extra whitespace",
+	async (props: {
+		currentVersion: SemanticVersionString
+		nextVersion: SemanticVersionString
+	}) => {
+		const originalContent = dedent`
+			{
+				"$schema": "https://json.schemastore.org/package.json",
+				"name": "@rainstormy/helix",
+				"version":     "${props.currentVersion}",
+				"type": "module",
+			}
+		`
+		const expectedPromotedContent = `${dedent`
+			{
+				"$schema": "https://json.schemastore.org/package.json",
+				"name": "@rainstormy/helix",
+				"version":     "${props.nextVersion}",
+				"type": "module",
+			}
+		`}\n`
+
+		const newRelease: Release = {
+			checks: [],
+			date: "2024-04-15",
+			version: props.nextVersion,
+		}
+		const actualPromotedContent = await promotePackageJson(
+			originalContent,
+			newRelease,
+		)
+
+		it(`updates the 'version' field to ${props.nextVersion} and preserves the preceding whitespace`, () => {
+			expect(actualPromotedContent).toBe(expectedPromotedContent)
+		})
+	},
+)
+
+describe.each`
+	currentVersion           | nextVersion
 	${"1.0.0"}               | ${"1.1.1"}
 	${"7.1.3-beta.7"}        | ${"7.1.3-beta.4"}
 	${"9.0.5-rc.0+3a1c790f"} | ${"9.0.4"}